Release 1.3.11 (tracks 200f407eaf28dd7856b6df42bf1721d5f72bff94)

+++ dagster-cloud/dagster_cloud/storage/schedules/queries.py
@@ -28,7 +28,7 @@ CREATE_JOB_TICK_MUTATION = """
 """
 
 UPDATE_JOB_TICK_MUTATION = """
-    mutation updateJobTickMutation($tickId: Int! $serializedJobTickData: String!) {
+    mutation updateJobTickMutation($tickId: BigInt! $serializedJobTickData: String!) {
         schedules {
             updateJobTick(tickId: $tickId, serializedJobTickData: $serializedJobTickData) {
                 ok
